+++ src/main/java/wooteco/subway/line/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line.controller;
+package wooteco.subway.line.dto;
 
 public class SectionRequest {
     private Long upStationId;
