+++ src/main/java/wooteco/subway/station/controller/StationController.java
@@ -6,9 +6,11 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
-import wooteco.subway.station.StationService;
 import wooteco.subway.station.StationValidator;
 import wooteco.subway.station.domain.Station;
+import wooteco.subway.station.dto.StationRequest;
+import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.service.StationService;
 
 import javax.validation.Valid;
 import java.net.URI;
