+++ src/test/java/wooteco/subway/line/LineAcceptanceTest.java
@@ -1,4 +1,5 @@
-package wooteco.subway.dto.station;
+package wooteco.subway.line;
+
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -10,91 +11,94 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
-import wooteco.subway.station.controller.StationResponse;
-import wooteco.subway.station.dao.StationDao;
+import wooteco.subway.line.dao.LineDao;
+import wooteco.subway.line.dto.LineResponse;
 
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("지하철역 관련 기능")
-public class StationAcceptanceTest extends AcceptanceTest {
+@DisplayName("노선 관련 기능")
+public class LineAcceptanceTest extends AcceptanceTest {
 
     @Autowired
-    private StationDao stationDao;
+    private LineDao lineDao;
+
+    private static final Map<String, String> params1 = new HashMap<>();
+    private static final Map<String, String> params2 = new HashMap<>();
+
+    static {
+        params1.put("name", "신분당선");
+        params1.put("color", "bg-red-600");
+        params2.put("name", "2호선");
+        params2.put("color", "bg-green-600");
+    }
 
     @AfterEach
     void clear() {
-        stationDao.clear();
+        lineDao.clear();
     }
 
-    @DisplayName("지하철역을 생성한다.")
+    @DisplayName("노선을 생성한다.")
     @Test
-    void createStation() {
-        // given
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "강남역");
-
+    void createLine() {
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .body(params)
+                .body(params1)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/stations")
+                .post("/lines")
                 .then().log().all()
                 .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).isNotBlank();
+        assertThat(response.body().jsonPath().getString("name")).isEqualTo("신분당선");
+        assertThat(response.body().jsonPath().getString("color")).isEqualTo("bg-red-600");
     }
 
-    @DisplayName("지하철 역 이름은 XX역으로 끝나지 않으면 예외처리가 된다.")
+    @DisplayName("지하철 노선 이름은 XX선으로 끝나지 않으면 예외처리가 된다.")
     @Test
-    void invalidStationName() {
+    void invalidLineName() {
         // given
         Map<String, String> invalidParam = new HashMap<>();
-        invalidParam.put("name", "역이름은역이라는단어로끝나야함");
+        invalidParam.put("name", "노선은선이라는단어로끝나야함");
 
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
                 .body(invalidParam)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/stations")
+                .post("/lines")
                 .then().log().all()
                 .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(response.body().asString()).isEqualTo("지하철 역 이름이 잘못되었습니다.");
+        assertThat(response.body().asString()).isEqualTo("지하철 노선 이름이 잘못되었습니다.");
     }
 
-    @DisplayName("중복된 지하철 역을 생성할 수 없다.")
+    @DisplayName("중복된 지하철 노선을 생성할 수 없다.")
     @Test
-    void cannotCreateDuplicatedStation() {
-        // given
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "강남역");
-
+    void cannotCreateDuplicatedLine() {
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .body(params)
+                .body(params1)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/stations")
+                .post("/lines")
                 .then().log().all()
                 .extract();
 
         ExtractableResponse<Response> response2 = RestAssured.given().log().all()
-                .body(params)
+                .body(params1)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/stations")
+                .post("/lines")
                 .then().log().all()
                 .extract();
 
@@ -103,98 +107,64 @@ public class StationAcceptanceTest extends AcceptanceTest {
         assertThat(response.header("Location")).isNotBlank();
 
         assertThat(response2.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(response2.body().asString()).isEqualTo("중복된 지하철 역입니다.");
+        assertThat(response2.body().asString()).isEqualTo("중복된 노선입니다.");
     }
 
-    @DisplayName("기존에 존재하는 지하철역 이름으로 지하철역을 생성한다.")
+    @DisplayName("노선을 조회한다.")
     @Test
-    void createStationWithDuplicateName() {
-        // given
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "강남역");
-        RestAssured.given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/stations")
-                .then().log().all()
-                .extract();
-
-        // when
-        ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/stations")
-                .then()
-                .log().all()
-                .extract();
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @DisplayName("지하철역을 조회한다.")
-    @Test
-    void showStations() {
+    void showLines() {
         /// given
-        Map<String, String> params1 = new HashMap<>();
-        params1.put("name", "강남역");
         ExtractableResponse<Response> createResponse1 = RestAssured.given().log().all()
                 .body(params1)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/stations")
+                .post("/lines")
                 .then().log().all()
                 .extract();
 
-        Map<String, String> params2 = new HashMap<>();
-        params2.put("name", "역삼역");
         ExtractableResponse<Response> createResponse2 = RestAssured.given().log().all()
                 .body(params2)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/stations")
+                .post("/lines")
                 .then().log().all()
                 .extract();
 
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
                 .when()
-                .get("/stations")
+                .get("/lines")
                 .then().log().all()
                 .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-        List<Long> expectedLineIds = Arrays.asList(createResponse1, createResponse2).stream()
-                .map(it -> Long.parseLong(it.header("Location").split("/")[2]))
+        List<Long> expectedLineIds = Stream.of(createResponse1, createResponse2)
+                .map(line -> Long.parseLong(line.header("Location").split("/")[2]))
                 .collect(Collectors.toList());
-        List<Long> resultLineIds = response.jsonPath().getList(".", StationResponse.class).stream()
-                .map(it -> it.getId())
+        List<Long> resultLineIds = response.jsonPath().getList(".", LineResponse.class).stream()
+                .map(line -> line.getId())
                 .collect(Collectors.toList());
         assertThat(resultLineIds).containsAll(expectedLineIds);
     }
 
-    @DisplayName("지하철역을 제거한다.")
+    @DisplayName("지하철 노선을 제거한다.")
     @Test
-    void deleteStation() {
+    void deleteLine() {
         // given
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "강남역");
         ExtractableResponse<Response> createResponse = RestAssured.given().log().all()
-                .body(params)
+                .body(params1)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/stations")
+                .post("/lines")
                 .then().log().all()
                 .extract();
 
         RestAssured.given().log().all()
-                .body("{\"name\": \"잠실역\"}")
+                .body(params2)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/stations")
+                .post("/lines")
                 .then().log().all()
                 .extract();
