+++ src/test/java/wooteco/subway/station/StationAcceptanceTest.java
@@ -1,5 +1,4 @@
-package wooteco.subway.dto.line;
-
+package wooteco.subway.station;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -11,94 +10,91 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
-import wooteco.subway.line.dao.LineDao;
-import wooteco.subway.line.controller.LineResponse;
+import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.dao.StationDao;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("노선 관련 기능")
-public class LineAcceptanceTest extends AcceptanceTest {
+@DisplayName("지하철역 관련 기능")
+public class StationAcceptanceTest extends AcceptanceTest {
 
     @Autowired
-    private LineDao lineDao;
-
-    private static final Map<String, String> params1 = new HashMap<>();
-    private static final Map<String, String> params2 = new HashMap<>();
-
-    static {
-        params1.put("name", "신분당선");
-        params1.put("color", "bg-red-600");
-        params2.put("name", "2호선");
-        params2.put("color", "bg-green-600");
-    }
+    private StationDao stationDao;
 
     @AfterEach
     void clear() {
-        lineDao.clear();
+        stationDao.clear();
     }
 
-    @DisplayName("노선을 생성한다.")
+    @DisplayName("지하철역을 생성한다.")
     @Test
-    void createLine() {
+    void createStation() {
+        // given
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "강남역");
+
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .body(params1)
+                .body(params)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/lines")
+                .post("/stations")
                 .then().log().all()
                 .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.body().jsonPath().getString("name")).isEqualTo("신분당선");
-        assertThat(response.body().jsonPath().getString("color")).isEqualTo("bg-red-600");
+        assertThat(response.header("Location")).isNotBlank();
     }
 
-    @DisplayName("지하철 노선 이름은 XX선으로 끝나지 않으면 예외처리가 된다.")
+    @DisplayName("지하철 역 이름은 XX역으로 끝나지 않으면 예외처리가 된다.")
     @Test
-    void invalidLineName() {
+    void invalidStationName() {
         // given
         Map<String, String> invalidParam = new HashMap<>();
-        invalidParam.put("name", "노선은선이라는단어로끝나야함");
+        invalidParam.put("name", "역이름은역이라는단어로끝나야함");
 
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
                 .body(invalidParam)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/lines")
+                .post("/stations")
                 .then().log().all()
                 .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(response.body().asString()).isEqualTo("지하철 노선 이름이 잘못되었습니다.");
+        assertThat(response.body().asString()).isEqualTo("지하철 역 이름이 잘못되었습니다.");
     }
 
-    @DisplayName("중복된 지하철 노선을 생성할 수 없다.")
+    @DisplayName("중복된 지하철 역을 생성할 수 없다.")
     @Test
-    void cannotCreateDuplicatedLine() {
+    void cannotCreateDuplicatedStation() {
+        // given
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "강남역");
+
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .body(params1)
+                .body(params)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/lines")
+                .post("/stations")
                 .then().log().all()
                 .extract();
 
         ExtractableResponse<Response> response2 = RestAssured.given().log().all()
-                .body(params1)
+                .body(params)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/lines")
+                .post("/stations")
                 .then().log().all()
                 .extract();
 
@@ -107,64 +103,98 @@ public class LineAcceptanceTest extends AcceptanceTest {
         assertThat(response.header("Location")).isNotBlank();
 
         assertThat(response2.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(response2.body().asString()).isEqualTo("중복된 노선입니다.");
+        assertThat(response2.body().asString()).isEqualTo("중복된 지하철 역입니다.");
     }
 
-    @DisplayName("노선을 조회한다.")
+    @DisplayName("기존에 존재하는 지하철역 이름으로 지하철역을 생성한다.")
     @Test
-    void showLines() {
+    void createStationWithDuplicateName() {
+        // given
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "강남역");
+        RestAssured.given().log().all()
+                .body(params)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .post("/stations")
+                .then().log().all()
+                .extract();
+
+        // when
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+                .body(params)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .post("/stations")
+                .then()
+                .log().all()
+                .extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+    }
+
+    @DisplayName("지하철역을 조회한다.")
+    @Test
+    void showStations() {
         /// given
+        Map<String, String> params1 = new HashMap<>();
+        params1.put("name", "강남역");
         ExtractableResponse<Response> createResponse1 = RestAssured.given().log().all()
                 .body(params1)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/lines")
+                .post("/stations")
                 .then().log().all()
                 .extract();
 
+        Map<String, String> params2 = new HashMap<>();
+        params2.put("name", "역삼역");
         ExtractableResponse<Response> createResponse2 = RestAssured.given().log().all()
                 .body(params2)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/lines")
+                .post("/stations")
                 .then().log().all()
                 .extract();
 
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
                 .when()
-                .get("/lines")
+                .get("/stations")
                 .then().log().all()
                 .extract();
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-        List<Long> expectedLineIds = Stream.of(createResponse1, createResponse2)
-                .map(line -> Long.parseLong(line.header("Location").split("/")[2]))
+        List<Long> expectedLineIds = Arrays.asList(createResponse1, createResponse2).stream()
+                .map(it -> Long.parseLong(it.header("Location").split("/")[2]))
                 .collect(Collectors.toList());
-        List<Long> resultLineIds = response.jsonPath().getList(".", LineResponse.class).stream()
-                .map(line -> line.getId())
+        List<Long> resultLineIds = response.jsonPath().getList(".", StationResponse.class).stream()
+                .map(it -> it.getId())
                 .collect(Collectors.toList());
         assertThat(resultLineIds).containsAll(expectedLineIds);
     }
 
-    @DisplayName("지하철 노선을 제거한다.")
+    @DisplayName("지하철역을 제거한다.")
     @Test
-    void deleteLine() {
+    void deleteStation() {
         // given
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "강남역");
         ExtractableResponse<Response> createResponse = RestAssured.given().log().all()
-                .body(params1)
+                .body(params)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/lines")
+                .post("/stations")
                 .then().log().all()
                 .extract();
 
         RestAssured.given().log().all()
-                .body(params2)
+                .body("{\"name\": \"잠실역\"}")
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
-                .post("/lines")
+                .post("/stations")
                 .then().log().all()
                 .extract();
